Added drawNames flag to street renderer, to easy activate rendering on names (still needs code change to activate it).
Added bool to indicate if a way type needs the outline, activated only for plain streets and squares and such.

Added command line parameters to osm2wai and wai itself, so you can now set which areas to render and what not, and output data to a specific directory, and load from it. You can 
also load several maps from wai. Defaults do the same as before: use the /tmp/ dir for writing and reading maps.